watch-screen-layout-changes: Remove broken duplicate event check
XRandR tends to emit two events for every screen layout change. This
code tried to catch this case to avoid unnecessary calls to the python
VM notification tool. Sadly, the check it did based on X server
timestamps did not work in practice. It did not catch duplicate events
in case of resolution or screen position changes but only in case of
screens removed, and it did spuriously catch *both* events in case of
screens added to the system.

I (jaseg) had a look at other possibilities to do this check, and there
does not seem to be a simple way purely from information contained in
the event structure. There are only two fixes I can think of right now:
 1) Fixing upstream (so XRandR only emits one event for each change)
 2) Rolling the xrandr-calling-and-parsing logic in the complimentary
    python script into a bunch of XRandR API calls in this C code and
    doing a simple equality check on both configurations as understood
    by qubes.
The former sounds like a lot of pain (and unlikely to land anywhere
usable in a reasonable amount of time) and for now I would like to avoid
proceeding with the latter as it sounds bug-prone at no great benefit.